feat: add scturcture for tracks
