+++ src/artists/interfaces/artist.interfaces.ts
