+++ src/tracks/tracks.controller.ts
@@ -1,0 +1,7 @@
+import { Controller } from '@nestjs/common';
+import { Response } from 'express';
+
+@Controller('track')
+export default class TracksController {
+  constructor(private readonly TracksService: TracksController) {}
+}
